requirements.txt
face_recognition
numpy
pillow
opencv-python